Refactor service restore flow into mixin to clear Codacy module-size gate
Extract restore/list-backup behavior from env_inspector_core/service.py into ServiceRestoreMixin and move target constants into env_inspector_core/targets.py. Keep compatibility behavior in EnvInspectorService and adjust test monkeypatch seam for the new restore module path.

Validation run locally:
- ruff check env_inspector_core/service.py env_inspector_core/service_restore.py env_inspector_core/targets.py env_inspector_core/wsl_targets.py tests/test_service_coverage_branches.py
- python -m pytest -q -s --cov=tests --cov-report=xml:coverage/python-coverage.xml
- python scripts/quality/assert_coverage_100.py --xml "python=coverage/python-coverage.xml" --min-percent 100
