+++ env_inspector_core/targets.py
@@ -1,0 +1,12 @@
+from __future__ import absolute_import, division
+
+TARGET_WINDOWS_USER = "windows:user"
+TARGET_WINDOWS_MACHINE = "windows:machine"
+TARGET_LINUX_BASHRC = "linux:bashrc"
+TARGET_LINUX_ETC_ENV = "linux:etc_environment"
+TARGET_POWERSHELL_CURRENT_USER = "powershell:current_user"
+TARGET_POWERSHELL_ALL_USERS = "powershell:all_users"
+DOTENV_TARGET_PREFIX = "dotenv:"
+WSL_DOTENV_TARGET_PREFIX = "wsl_dotenv:"
+WSL_DOTENV_PATH_ERROR = "Unsupported WSL dotenv target path"
+LINUX_ETC_ENV_PATH = "/etc/environment"
